Préparation à la suite (correction de petit warning). Le désassemblage dynamique risque d'être coton (comment le représenter ?)
git-svn-id: svn+ssh://svn.code.sf.net/p/desassembleur/code/trunk@45 eb032535-49dc-4373-a4b2-a5adefcad844

+++ dyndesass.c
@@ -41,7 +41,7 @@ void desassemblage_dynamique(DISASM* prog) {
     int len;
 
     while (!erreur) {
-        prog->SecurityBlock = finProg - prog->EIP;
+        prog->SecurityBlock = (unsigned int) (finProg - prog->EIP);
         len = Disasm(prog);
         //printf("SecurityBlock = %d et len = %d \n", prog->SecurityBlock, len);
         if (len == UNKNOWN_OPCODE) {
